+++ CHANGELOG.md
@@ -241,6 +241,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
   - From\<&[u8; N]\> for Contents
   - From\<&str\> for Content
   - From\<Vec\<u8\>\> for Content
+  - From\<String\> for Content
   - From\<&[u8; N]\> for &ContentsRef
   - From\<&mut [u8; N]\> for &mut ContentsRef
   - From\<&str\> for &ContentsRef